refactor(sensor): #74
